+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- webpack compiling for dustpress.js
+- js sourcemaps
+
 ## [3.0.4] - 2018-08-09
 
 ### Fixed